Controlerをテストように変更
